+++ packages/core/src/storages/request_queue.ts
@@ -9,14 +9,7 @@ import type {
     RequestQueueBackend,
     RequestQueueInfo,
 } from '@crawlee/types';
-import {
-    chunkedAsyncIterable,
-    downloadListOfUrls,
-    isAsyncIterable,
-    isIterable,
-    peekableAsyncIterable,
-    sleep,
-} from '@crawlee/utils';
+import { downloadListOfUrls, sleep } from '@crawlee/utils';
 import ow from 'ow';
 import type { ReadonlyDeep } from 'type-fest';
 
@@ -27,6 +20,7 @@ import { getObjectType } from '../debug.js';
 import type { Constructor } from '../typedefs.js';
 import type { EventManager } from '../events/event_manager.js';
 import { EventType } from '../events/event_manager.js';
+import { chunkedAsyncIterable, isAsyncIterable, isIterable, peekableAsyncIterable } from '../iterables.js';
 import type { CrawleeLogger } from '../log.js';
 import type { IProxyConfiguration } from '../proxy_configuration.js';
 import type { InternalSource, RequestOptions, Source } from '../request.js';
